+++ src/tilesetProcessing/upgrade/TilesetObjectUpgrader.ts
@@ -53,6 +53,10 @@ export class TilesetObjectUpgrader {
       logger.debug(`Upgrading content.url to content.uri`);
       await this.upgradeEachContentUrlToUri(tileset);
     }
+    if (this.upgradeOptions.upgradeEmptyChildrenToUndefined) {
+      logger.debug(`Upgrading empty children arrays to be 'undefined'`);
+      await this.upgradeEmptyChildrenToUndefined(tileset);
+    }
     if (this.upgradeOptions.upgradeContentGltfExtensionDeclarations) {
       logger.debug(`Upgrading extension declarations`);
       Extensions.removeExtensionUsed(tileset, "3DTILES_content_gltf");
@@ -125,6 +129,23 @@ export class TilesetObjectUpgrader {
     logger.warn(
       "  The content does not have a 'uri' property (and no legacy 'url' property)"
     );
+  }
+
+  /**
+   * Upgrade each empty 'children' in any tile by deleting them and
+   * causing the children to become 'undefined'
+   *
+   * @param tileset - The tileset
+   */
+  private async upgradeEmptyChildrenToUndefined(tileset: Tileset) {
+    const root = tileset.root;
+    await Tiles.traverseExplicit(root, async (tilePath: Tile[]) => {
+      const tile = tilePath[tilePath.length - 1];
+      if (tile.children !== undefined && tile.children.length === 0) {
+        delete tile.children;
+      }
+      return true;
+    });
   }
 
   /**